Eliminate non 3rd party related warnings from daemon

+++ src/AppDaemon.cpp
@@ -121,7 +121,7 @@ bool AppDaemon::initialize()
     if (parser.isSet(debugHttpServer))
     {
         httpServer = new HttpServer(this);
-        if (!httpServer->start(parser.value(debugHttpServer).toInt()))
+        if (!httpServer->start(static_cast<quint16>(parser.value(debugHttpServer).toInt())))
         {
             qCritical() << "Fatal error: Failed to create HTTP server.";
             return false;
